+++ pages/_app.tsx
@@ -5,9 +5,12 @@ import { DefaultSeo } from "next-seo";
 import "tailwindcss/tailwind.css";
 import "@styles/global.scss";
 import "highlight.js/styles/stackoverflow-dark.css";
+import { Header } from "@components/header";
+import { Footer } from "@components/footer";
 
 function MyApp({ Component, pageProps, router }: AppProps): JSX.Element {
     const url = `https://localhost:3000${router.route}`;
+
     return (
         <>
             <DefaultSeo
@@ -23,7 +26,9 @@ function MyApp({ Component, pageProps, router }: AppProps): JSX.Element {
                 }}
                 canonical={url}
             />
+            <Header />
             <Component {...pageProps} />
+            <Footer />
         </>
     );
 }