+++ pages/blog/index.tsx
@@ -1,6 +1,5 @@
 import { Layout } from "@components/layout";
 import { PostPanel } from "@components/panel";
-import { Parallax } from "@components/parallax";
 import { Section } from "@components/section";
 import { IPost } from "@util/interface";
 import { getBlogPosts } from "@util/notion";
@@ -15,7 +14,7 @@ const BlogPage: NextPage<BlockPageProps> = ({ posts }: BlockPageProps) => {
         <Layout title="Blog" description="My engineering blog">
             <Section>
                 <div className="blog-section">
-                    <Parallax range={[0, -0.5]} className="blog-hero">
+                    <div className="blog-hero">
                         <span className="blog-title">My Blog</span>
                         <div className="introduction">
                             <p>
@@ -30,7 +29,7 @@ const BlogPage: NextPage<BlockPageProps> = ({ posts }: BlockPageProps) => {
                                 pulvinar quis diam non tempor. <br />
                             </p>
                         </div>
-                    </Parallax>
+                    </div>
                     <div className="blog-main">
                         <PostPanel posts={posts} />
                     </div>