+++ src/components/header/index.tsx
@@ -1,5 +1,6 @@
 import { PrimaryButton } from "@components/button";
 import { Menu } from "@components/menu";
+import { isActiveLink } from "@util/router";
 import {
     motion,
     useSpring,
@@ -8,24 +9,28 @@ import {
     useViewportScroll,
 } from "framer-motion";
 import Link from "next/link";
+import { useRouter } from "next/router";
 import React from "react";
 import { FaArrowUp } from "react-icons/fa";
 
 export const Header: React.FC = () => {
-    const { scrollYProgress } = useViewportScroll();
+    const { pathname } = useRouter();
+    const isHome = isActiveLink("/", pathname);
+
+    const { scrollY } = useViewportScroll();
     const buttonAppearance = useTransform(
-        scrollYProgress,
-        [0, 0.3, 0.35],
+        scrollY,
+        [0, 200, 250],
         [-100, -100, 0],
     );
     const headerDisappearance = useTransform(
-        scrollYProgress,
-        [0, 0.25, 0.3],
+        scrollY,
+        [0, 175, 225],
         [0, 0, -300],
     );
     const time = useTime();
     const headerAppearance = useSpring(
-        useTransform(time, [0, 1500, 1750], [0, 0, 1]),
+        useTransform(time, [0, 1500, 1750], isHome ? [0, 0, 1] : [1, 1, 1]),
         { stiffness: 300 },
     );
     return (